add: constants in the i18n. add: settings panel hooks
