+++ src/components/video-editor/settings/constants.ts
@@ -1,0 +1,85 @@
+import type { AutoCaptionAnimation, CursorStyle, WebcamPositionPreset, ZoomDepth } from "../types";
+
+export type CursorStyleOption = { value: CursorStyle; label: string };
+
+export const BUILTIN_CURSOR_PREVIEW_SIZE = 28;
+export const BUILTIN_CURSOR_PREVIEW_FRAME_SIZE = 48;
+
+export const GRADIENTS = [
+	"linear-gradient( 111.6deg,  rgba(114,167,232,1) 9.4%, rgba(253,129,82,1) 43.9%, rgba(253,129,82,1) 54.8%, rgba(249,202,86,1) 86.3% )",
+	"linear-gradient(120deg, #d4fc79 0%, #96e6a1 100%)",
+	"radial-gradient( circle farthest-corner at 3.2% 49.6%,  rgba(80,12,139,0.87) 0%, rgba(161,10,144,0.72) 83.6% )",
+	"linear-gradient( 111.6deg,  rgba(0,56,68,1) 0%, rgba(163,217,185,1) 51.5%, rgba(231, 148, 6, 1) 88.6% )",
+	"linear-gradient( 107.7deg,  rgba(235,230,44,0.55) 8.4%, rgba(252,152,15,1) 90.3% )",
+	"linear-gradient( 91deg,  rgba(72,154,78,1) 5.2%, rgba(251,206,70,1) 95.9% )",
+	"radial-gradient( circle farthest-corner at 10% 20%,  rgba(2,37,78,1) 0%, rgba(4,56,126,1) 19.7%, rgba(85,245,221,1) 100.2% )",
+	"linear-gradient( 109.6deg,  rgba(15,2,2,1) 11.2%, rgba(36,163,190,1) 91.1% )",
+	"linear-gradient(135deg, #FBC8B4, #2447B1)",
+	"linear-gradient(109.6deg, #F635A6, #36D860)",
+	"linear-gradient(90deg, #FF0101, #4DFF01)",
+	"linear-gradient(315deg, #EC0101, #5044A9)",
+	"linear-gradient(45deg, #ff9a9e 0%, #fad0c4 99%, #fad0c4 100%)",
+	"linear-gradient(to top, #a18cd1 0%, #fbc2eb 100%)",
+	"linear-gradient(to right, #ff8177 0%, #ff867a 0%, #ff8c7f 21%, #f99185 52%, #cf556c 78%, #b12a5b 100%)",
+	"linear-gradient(120deg, #84fab0 0%, #8fd3f4 100%)",
+	"linear-gradient(to right, #4facfe 0%, #00f2fe 100%)",
+	"linear-gradient(to top, #fcc5e4 0%, #fda34b 15%, #ff7882 35%, #c8699e 52%, #7046aa 71%, #0c1db8 87%, #020f75 100%)",
+	"linear-gradient(to right, #fa709a 0%, #fee140 100%)",
+	"linear-gradient(to top, #30cfd0 0%, #330867 100%)",
+	"linear-gradient(to top, #c471f5 0%, #fa71cd 100%)",
+	"linear-gradient(to right, #f78ca0 0%, #f9748f 19%, #fd868c 60%, #fe9a8b 100%)",
+	"linear-gradient(to top, #48c6ef 0%, #6f86d6 100%)",
+	"linear-gradient(to right, #0acffe 0%, #495aff 100%)",
+] as const;
+
+export const CAPTION_ANIMATION_OPTIONS: Array<{ value: AutoCaptionAnimation; label: string }> = [
+	{ value: "none", label: "Off" },
+	{ value: "fade", label: "Fade" },
+	{ value: "rise", label: "Rise" },
+	{ value: "pop", label: "Pop" },
+];
+
+export const ZOOM_DEPTH_OPTIONS: Array<{ depth: ZoomDepth; label: string }> = [
+	{ depth: 1, label: "1.25×" },
+	{ depth: 2, label: "1.5×" },
+	{ depth: 3, label: "1.8×" },
+	{ depth: 4, label: "2.2×" },
+	{ depth: 5, label: "3.5×" },
+	{ depth: 6, label: "5×" },
+];
+
+export const WEBCAM_POSITION_PRESETS: Array<{
+	preset: Exclude<WebcamPositionPreset, "custom">;
+	label: string;
+}> = [
+	{ preset: "top-left", label: "↖" },
+	{ preset: "top-center", label: "↑" },
+	{ preset: "top-right", label: "↗" },
+	{ preset: "center-left", label: "←" },
+	{ preset: "center", label: "•" },
+	{ preset: "center-right", label: "→" },
+	{ preset: "bottom-left", label: "↙" },
+	{ preset: "bottom-center", label: "↓" },
+	{ preset: "bottom-right", label: "↘" },
+];
+
+export const BUILTIN_CURSOR_STYLE_OPTIONS: CursorStyleOption[] = [
+	{ value: "macos", label: "macOS" },
+	{ value: "tahoe", label: "Tahoe" },
+	{ value: "tahoe-inverted", label: "Tahoe Inverted" },
+	{ value: "dot", label: "Dot" },
+	{ value: "figma", label: "Minimal" },
+];
+
+export const CAPTION_LANGUAGE_OPTIONS = [
+	{ value: "auto", label: "Auto Detect" },
+	{ value: "en", label: "English" },
+	{ value: "es", label: "Spanish" },
+	{ value: "fr", label: "French" },
+	{ value: "de", label: "German" },
+	{ value: "it", label: "Italian" },
+	{ value: "pt", label: "Portuguese" },
+	{ value: "zh", label: "Chinese (Simplified)" },
+	{ value: "ja", label: "Japanese" },
+	{ value: "ko", label: "Korean" },
+] as const;
