+++ src/i18n/config.ts
@@ -15,3 +15,15 @@ export const I18N_NAMESPACES = [
 
 export type AppLocale = (typeof SUPPORTED_LOCALES)[number];
 export type I18nNamespace = (typeof I18N_NAMESPACES)[number];
+
+export const APP_LANGUAGE_LABELS: Record<AppLocale, string> = {
+	en: "English",
+	es: "Español",
+	fr: "Français",
+	it: "Italiano",
+	nl: "Nederlands",
+	ko: "한국어",
+	"pt-BR": "Português",
+	"zh-CN": "簡體中文",
+	"zh-TW": "繁體中文",
+};
